+++ blueprints/ember-flexberry-data/index.js
@@ -9,7 +9,8 @@ module.exports = {
     }).then(function () {
       return _this.addPackagesToProject([
         { name: 'dexie', target: '2.0.2' },
-        { name: 'node-uuid', target: '1.4.7' }
+        { name: 'node-uuid', target: '1.4.7' },
+        { name: 'ember-moment', target: '7.6.0' }
       ]);
     });
   },